requirements.txt
pymysql
pandas
streamlit
plotly
mysql-connector-python
schedule
